requirements.txt
apache-airflow-providers-apache-spark==3.0.0
apache-airflow-providers-postgres==5.10.0
minio
lxml
pandas
pyarrow
tomli
requests
beautifulsoup4
trino
psycopg2-binary
rich<14
scrapy
pyarrow
protobuf>=5.0,<6.0
asyncpg
pyarrow
pandas
vn_address_converter
vietnamadminunits